% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{fix_study_id}
\alias{fix_study_id}
\title{Clean up a study ID that may contain non-standard characters}
\usage{
fix_study_id(x)
}
\arguments{
\item{x}{a character or character vector to "fix"}
}
\description{
Clean up a study ID that may contain non-standard characters
}
\details{
This simply converts any non-alphanumeric character to "_", which is a useful conversion for file names, etc.
}
