% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_study_list}
\alias{get_study_list}
\title{Get a data frame of meta data about all "hbgd" studies}
\usage{
get_study_list()
}
\description{
Get a data frame of meta data about all "hbgd" studies
}
\note{
Only studies are returned that have non-NA grant_folder
}
\examples{
\dontrun{
studies <- get_study_list()
wsb <- use_study("wsb")
}
}

