% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_study_list_anthro}
\alias{get_study_list_anthro}
\title{Get a data frame of meta data about studies to be used}
\usage{
get_study_list_anthro()
}
\description{
Get a data frame of meta data about studies to be used
}
\note{
Unlike \code{\link{get_study_list}}, this function returns a subset of studies that are guaranteed to return data in a standard format that includes anthropometry when used with \code{\link{use_study}}. Also note that there very well may be studies in the larger list returned by \code{\link{get_study_list}} that have the same kind of structure too but for some reason are not included in the results of this function.
}
\examples{
\dontrun{
studies <- get_study_list_anthro()
wsb <- use_study("wsb")
}
}
