% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdtm_utilities.R
\name{get_tbls}
\alias{get_tbls}
\title{Parse docx tables to list of data.frames}
\usage{
get_tbls(word_doc)
}
\arguments{
\item{word_doc}{character, path to word docx file}
}
\value{
list
}
\description{
Takes a docx as an imput and returns list of data.frames of all tables in document
}
