% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_adhocs_setup}
\alias{git_adhocs_setup}
\title{Check out the Vis-AdHocs repository in the correct location in the git directory}
\usage{
git_adhocs_setup()
}
\description{
Check out the Vis-AdHocs repository in the correct location in the git directory
}

