% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_setup}
\alias{git_setup}
\title{Find the local base git path and clone it if it isn't there}
\usage{
git_setup()
}
\description{
Find the local base git path and clone it if it isn't there
}
\note{
This function is usually called by other functions so it will rarely have need to be called explicitly.
}
