% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_ghap-data.R
\docType{data}
\name{meta_ghap}
\alias{meta_ghap}
\title{Meta data that search_ghap is run against.}
\format{A data frame with 9209 rows and 23 variables:
\describe{
  \item{\code{DOMAIN}}{character Domain of study}
  \item{\code{STUDYID}}{character Study ID}
  \item{\code{variable}}{character Variable in Study ID}
  \item{\code{LABEL}}{character Label of variable}
  \item{\code{DOMAIN.LBL}}{character Domain Label}
  \item{\code{DOMAIN.STRUCTURE}}{character Domain Structure}
  \item{\code{Study_ID}}{character Study Id}
  \item{\code{Short_Description}}{character Short description of Study}
  \item{\code{Study_Description}}{character Long description of Study}
  \item{\code{Alternate_ID}}{character Alternative ID of Study}
  \item{\code{Subject_Count}}{integer Number of Subjects in Study}
  \item{\code{Country}}{character Countries that study include}
  \item{\code{Data_Status}}{character QA status of Data}
  \item{\code{Study_Type}}{character Longitudinal or Cross sectional}
  \item{\code{Intervention_Type}}{character Intervention Type}
  \item{\code{Age_Lower_Limit}}{character Lower bound of Age in Study}
  \item{\code{Age_Upper_Limit}}{character Upper bound of Age in Study}
  \item{\code{Start_Year}}{integer Start year of Study}
  \item{\code{Stop_Year}}{integer End year of Study}
  \item{\code{Population}}{character Population of Study}
  \item{\code{Repository_Name}}{character Name of repository that study is in}
  \item{\code{Repository_Subfolder}}{character Subdirectory study is located in}
  \item{\code{Study_URL}}{character Study URL}
}}
\usage{
meta_ghap
}
\description{
Meta data that is built from cross_sectional and longitudinal
 repositories on ghap
}
\keyword{datasets}
