% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdtm_utilities.R
\name{parse_docs}
\alias{parse_docs}
\title{Create sdtm structure files from ASDTM tables from docx file}
\usage{
parse_docs(path)
}
\arguments{
\item{path}{character, path to word docx file}
}
\value{
list
}
\description{
Create files that summarize sdtm contents, files and domain variable meta information
}
