% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{paste_repos}
\alias{paste_repos}
\title{Retrieve git hbgd repository addresses from data store explorer}
\usage{
paste_repos()
}
\value{
data.frame
}
\description{
Copy URL of current data store explorer state on clipboard and function returns list of
checked studies with their respective clone paths on the git repository.
}

