% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_and_upload}
\alias{save_and_upload}
\title{Save a plot and upload it to a project on OSF (osf.io)}
\usage{
save_and_upload(p, filename, width = 800, height = 800, type = "png",
  osf_id, osf_folder = "plots", ...)
}
\arguments{
\item{p}{specification (lattice object, ggplot2 object, or expression of arbitrary R code)}

\item{filename}{where to save the file}

\item{width}{(in pixels)}

\item{height}{(in pixels)}

\item{type}{type of file ("png" or "pdf")}

\item{osf_id}{id of the project on OSF (osf.io/xxxx)}

\item{osf_folder}{folder within the project to save the file (file will have same base name as \code{filename})}

\item{\ldots}{additional parameters passed to graphics device}
}
\description{
Save a plot and upload it to a project on OSF (osf.io)
}
\details{
All GHAP repositories are automatically organized and stored in a directory structure under a base git path which is specified by this function. This function will set an R environment variable that will be present in all subsequent sessions such that you only need to specify the base path once.
}
