% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ghap.R
\name{search_ghap}
\alias{search_ghap}
\title{api to search ghap respositories}
\usage{
search_ghap(query, field = "LABEL", complete = FALSE, ...)
}
\arguments{
\item{query}{character, search expression}

\item{field}{character, column of meta data to search on, Default: 'LABEL'}

\item{complete}{boolean, only show complete cases of query result, Default: FALSE}

\item{...}{parameters to pass to grepl}
}
\value{
data.frame
}
\description{
use regular expression to search fields in ghap study meta data
}
\examples{
search_ghap(c('^Birth','birth \\\\(days\\\\)$'),complete = FALSE)
}
