% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbuilder.R
\name{searchbuilder}
\alias{searchbuilder}
\title{Shinyapp to navigate and maintain GHAP repositories}
\usage{
searchbuilder(viewer = shiny::dialogViewer(dialogName = "GHAP", width = 3000,
  height = 2000))
}
\arguments{
\item{viewer}{Where to open the application can be dialogViewer, browserViewer or paneViewer , Default: shiny::dialogViewer()}
}
\value{
nothing
}
\description{
Shinyapp to navigate and maintain GHAP repositories run from console or launch from addin menu
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 searchbuilder()
 }
}
}
