% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{set_git_base_path}
\alias{set_git_base_path}
\title{Set the base path for all repositories checked out from git}
\usage{
set_git_base_path(path)
}
\arguments{
\item{path}{a directory in which to place all checked out repositories (should be an empty directory)}
}
\description{
Set the base path for all repositories checked out from git
}
\details{
All GHAP repositories are automatically organized and stored in a directory structure under a base git path which is specified by this function. This function will set an R environment variable that will be present in all subsequent sessions such that you only need to specify the base path once.
}

