% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{set_git_credentials}
\alias{set_git_credentials}
\title{Set git credentials}
\usage{
set_git_credentials()
}
\description{
Set git credentials
}
\note{
This function is usually called by other functions so it will rarely have need to be called explicitly.
}
