% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_ghap.R
\name{sparse_ghap}
\alias{sparse_ghap}
\title{Create a sparse ghap repository}
\usage{
sparse_ghap(repo_url, repo, dirs, create = TRUE, remote = "origin",
  branch = "master")
}
\arguments{
\item{repo_url}{character, repository url path}

\item{repo}{character, path on local disk that repository is cloned to}

\item{dirs}{character, vector of repository subdirectories to fetch}

\item{create}{boolean, create a new git clone?, Default: TRUE}

\item{remote}{character, alias of the remote Default: 'origin'}

\item{branch}{character, alias of the branch, Default: 'master'}
}
\value{
nothing
}
\description{
Create a ghap repository that fetches only specified subdirectories and file types
}
