% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_ghap-data.R
\docType{data}
\name{meta_ghap}
\alias{meta_ghap}
\title{Meta data that search_ghap is run against.}
\format{A data frame with 9209 rows and 23 variables:
\describe{
  \item{\code{DOMAIN}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_VARIABLE}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_VARIABLE_DESCRIPTION}}{character COLUMN_DESCRIPTION}
  \item{\code{DOMAIN_DESCRIPTION}}{character COLUMN_DESCRIPTION}
  \item{\code{DOMAIN_STRUCTURE}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_ID_SHORT}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_DESCRIPTION_SHORT}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_DESCRIPTION}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_ID_ALTERNATE}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_SUBJECT_COUNT}}{integer COLUMN_DESCRIPTION}
  \item{\code{STUDY_COUNTRY}}{character COLUMN_DESCRIPTION}
  \item{\code{REPOSITORY_DATA_STATUS}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_TYPE}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_INTERVENTION_TYPE}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_AGE_LOWER_LIMIT}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_AGE_UPPER_LIMIT}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_START_YEAR}}{integer COLUMN_DESCRIPTION}
  \item{\code{STUDY_STOP_YEAR}}{integer COLUMN_DESCRIPTION}
  \item{\code{STUDY_POPULATION}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_REPOSITORY_NAME}}{character COLUMN_DESCRIPTION}
  \item{\code{REPOSITORY_SUBFOLDER}}{character COLUMN_DESCRIPTION}
  \item{\code{STUDY_URL}}{character COLUMN_DESCRIPTION} 
}}
\usage{
meta_ghap
}
\description{
Meta data that is built from cross_sectional and longitudinal
 repositories on ghap
}
\keyword{datasets}
