% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{use_study}
\alias{use_study}
\title{Get data for a given study ID}
\usage{
use_study(id)
}
\arguments{
\item{id}{the study ID - this can either be a short ID or long ID which are found in the ID columns of the result of \code{\link{get_study_list}} or \code{\link{get_study_list_anthro}}}
}
\description{
Get data for a given study ID
}
\details{
This function takes a study ID and if there is not a git repository checked out for it, it will check out the respository and read and return the appropriate data file. If the repository is checked out, it will pull any updates to the data and then read and return the appropriate data file. In the case of study IDs associated with \code{\link{get_study_list_anthro}}, the correct data will be returned. In the case of study IDs not in this list but in \code{\link{get_study_list}}, a guess will be made as to which data file is appropriate.
}
\examples{
\dontrun{
studies <- get_study_list_anthro()
wsb <- use_study("wsb")
}
}
